server/requirements.txt
fastapi
uvicorn
pdfplumber
PyMuPDF
python-dotenv
requests
anthropic
python-multipart
supabase
